N Division - L1
Write a program that will find all such numbers which are divisible by 7 but are not a multiple of 5. The numbers obtained should be printed in a comma-separated sequence on a single line.

Input Format:
The first line consists of an integer value as starting value
The second line consists of an integer value as the ending value.

Output Format:
The output consists of numbers that are all divisible by 7 and are not a multiple of 5 with the given range.